Remove singleton pattern
