+++ Client/Frontend/Theme/DarkTheme.swift
@@ -13,34 +13,54 @@ struct DarkTheme: Theme {
 }
 
 private struct DarkColourPalette: ThemeColourPalette {
-// TODO: Replace with colors from the Mobile Styles Figma
-    var actionPrimary: UIColor { return UIColor(named: "ActionPrimary")! }
-    var actionSecondary: UIColor { return UIColor(named: "ActionSecondary")! }
-    var borderDivider: UIColor { return UIColor(named: "BorderDivider")! }
-    var borderSelectedDefault: UIColor { return UIColor(named: "BorderSelectedDefault")! }
-    var borderSelectedPrivate: UIColor { return UIColor(named: "BorderSelectedPrivate")! }
-    var controlActive: UIColor { return UIColor(named: "ControlActive")! }
-    var controlBase: UIColor { return UIColor(named: "ControlBase")! }
-    var iconAccentBlue: UIColor { return UIColor(named: "IconAccentBlue")! }
-    var iconAccentGreen: UIColor { return UIColor(named: "IconAccentGreen")! }
-    var iconAccentPink: UIColor { return UIColor(named: "IconAccentPink")! }
-    var iconAccentViolet: UIColor { return UIColor(named: "IconAccentViolet")! }
-    var iconAccentYellow: UIColor { return UIColor(named: "IconAccentYellow")! }
-    var iconDisabled: UIColor { return UIColor(named: "IconDisabled")! }
-    var iconInverted: UIColor { return UIColor(named: "IconInverted")! }
-    var iconPrimary: UIColor { return UIColor(named: "IconPrimary")! }
-    var iconSecondary: UIColor { return UIColor(named: "IconSecondary")! }
-    var layer1: UIColor { return UIColor(named: "Layer1")! }
-    var layer2: UIColor { return UIColor(named: "Layer2")! }
-    var layer2Blur: UIColor { return UIColor(named: "Layer2Blur")! }
-    var layer3: UIColor { return UIColor(named: "Layer3")! }
-    var layer4: UIColor { return UIColor(named: "Layer4")! }
-    var layerEmphasis: UIColor { return UIColor(named: "LayerEmphasis")! }
-    var scrim: UIColor { return UIColor(named: "Scrim")! }
-    var textDisabled: UIColor { return UIColor(named: "TextDisabled")! }
-    var textInverted: UIColor { return UIColor(named: "TextInverted")! }
-    var textLink: UIColor { return UIColor(named: "TextLink")! }
-    var textPrimary: UIColor { return UIColor(named: "TextPrimary")! }
-    var textSecondary: UIColor { return UIColor(named: "TextSecondary")! }
-    var textWarning: UIColor { return UIColor(named: "TextWarning")! }
+    // MARK: - Layers
+    var layer1: UIColor = FXColors.DarkGrey60
+    var layer2: UIColor = FXColors.DarkGrey30
+    var layer3: UIColor = FXColors.DarkGrey80
+    var layer4: UIColor = FXColors.DarkGrey20.withAlphaComponent(0.7)
+    var layer5: UIColor = FXColors.DarkGrey40
+    var layerScrim: UIColor = FXColors.DarkGrey90.withAlphaComponent(0.95)
+    var layerGradient: Gradient = Gradient(start: FXColors.Violet70, end: FXColors.Violet40)
+    var layerAccentNonOpaque: UIColor = FXColors.Blue20.withAlphaComponent(0.2)
+    var layerAccentPrivate: UIColor = FXColors.Purple60
+    var layerAccentPrivateNonOpaque: UIColor = FXColors.Purple60.withAlphaComponent(0.3)
+    var layerLightGrey30: UIColor = FXColors.LightGrey30
+
+    // MARK: - Actions
+    var actionPrimary: UIColor = FXColors.Blue20
+    var actionPrimaryHover: UIColor = FXColors.Blue10
+    var actionSecondary: UIColor = FXColors.LightGrey30
+    var actionSecondaryHover: UIColor = FXColors.LightGrey20
+    var formSurfaceOff: UIColor = FXColors.DarkGrey05
+    var formKnob: UIColor = FXColors.White
+    var indicatorActive: UIColor = FXColors.LightGrey90
+    var indicatorInactive: UIColor = FXColors.DarkGrey05
+
+    // MARK: - Text
+    var textPrimary: UIColor = FXColors.LightGrey05
+    var textSecondary: UIColor = FXColors.LightGrey40
+    var textDisabled: UIColor = FXColors.LightGrey05.withAlphaComponent(0.4)
+    var textWarning: UIColor = FXColors.Red20
+    var textAccent: UIColor = FXColors.Blue20
+    var textOnColor: UIColor = FXColors.LightGrey05
+    var textInverted: UIColor = FXColors.DarkGrey90
+
+    // MARK: - Icons
+    var iconPrimary: UIColor = FXColors.LightGrey05
+    var iconSecondary: UIColor = FXColors.LightGrey40
+    var iconDisabled: UIColor = FXColors.LightGrey05.withAlphaComponent(0.4)
+    var iconAction: UIColor = FXColors.Blue20
+    var iconOnColor: UIColor = FXColors.LightGrey05
+    var iconWarning: UIColor = FXColors.Red20
+    var iconAccentViolet: UIColor = FXColors.Violet20
+    var iconAccentBlue: UIColor = FXColors.Blue20
+    var iconAccentPink: UIColor = FXColors.Pink20
+    var iconAccentGreen: UIColor = FXColors.Green20
+    var iconAccentYellow: UIColor = FXColors.Yellow20
+
+    // MARK: - Border
+    var borderPrimary: UIColor = FXColors.DarkGrey05
+    var borderAccent: UIColor = FXColors.Blue20
+    var borderAccentNonOpaque: UIColor = FXColors.Blue20.withAlphaComponent(0.2)
+    var borderAccentPrivate: UIColor = FXColors.Purple60
 }
