+++ Client/Frontend/Theme/LightTheme.swift
@@ -14,34 +14,54 @@ struct LightTheme: Theme {
 }
 
 private struct LightColourPalette: ThemeColourPalette {
-// TODO: Replace with colors from the Mobile Styles Figma
-    var actionPrimary: UIColor { return UIColor(named: "ActionPrimary")! }
-    var actionSecondary: UIColor { return UIColor(named: "ActionSecondary")! }
-    var borderDivider: UIColor { return UIColor(named: "BorderDivider")! }
-    var borderSelectedDefault: UIColor { return UIColor(named: "BorderSelectedDefault")! }
-    var borderSelectedPrivate: UIColor { return UIColor(named: "BorderSelectedPrivate")! }
-    var controlActive: UIColor { return UIColor(named: "ControlActive")! }
-    var controlBase: UIColor { return UIColor(named: "ControlBase")! }
-    var iconAccentBlue: UIColor { return UIColor(named: "IconAccentBlue")! }
-    var iconAccentGreen: UIColor { return UIColor(named: "IconAccentGreen")! }
-    var iconAccentPink: UIColor { return UIColor(named: "IconAccentPink")! }
-    var iconAccentViolet: UIColor { return UIColor(named: "IconAccentViolet")! }
-    var iconAccentYellow: UIColor { return UIColor(named: "IconAccentYellow")! }
-    var iconDisabled: UIColor { return UIColor(named: "IconDisabled")! }
-    var iconInverted: UIColor { return UIColor(named: "IconInverted")! }
-    var iconPrimary: UIColor { return UIColor(named: "IconPrimary")! }
-    var iconSecondary: UIColor { return UIColor(named: "IconSecondary")! }
-    var layer1: UIColor { return UIColor(named: "Layer1")! }
-    var layer2: UIColor { return UIColor(named: "Layer2")! }
-    var layer2Blur: UIColor { return UIColor(named: "Layer2Blur")! }
-    var layer3: UIColor { return UIColor(named: "Layer3")! }
-    var layer4: UIColor { return UIColor(named: "Layer4")! }
-    var layerEmphasis: UIColor { return UIColor(named: "LayerEmphasis")! }
-    var scrim: UIColor { return UIColor(named: "Scrim")! }
-    var textDisabled: UIColor { return UIColor(named: "TextDisabled")! }
-    var textInverted: UIColor { return UIColor(named: "TextInverted")! }
-    var textLink: UIColor { return UIColor(named: "TextLink")! }
-    var textPrimary: UIColor { return UIColor(named: "TextPrimary")! }
-    var textSecondary: UIColor { return UIColor(named: "TextSecondary")! }
-    var textWarning: UIColor { return UIColor(named: "TextWarning")! }
+    // MARK: - Layers
+    var layer1: UIColor = FXColors.LightGrey10
+    var layer2: UIColor = FXColors.White
+    var layer3: UIColor = FXColors.LightGrey20
+    var layer4: UIColor = FXColors.LightGrey30.withAlphaComponent(0.6)
+    var layer5: UIColor = FXColors.White
+    var layerScrim: UIColor = FXColors.DarkGrey30.withAlphaComponent(0.95)
+    var layerGradient: Gradient = Gradient(start: FXColors.Violet70, end: FXColors.Violet40)
+    var layerAccentNonOpaque: UIColor = FXColors.Blue50.withAlphaComponent(0.1)
+    var layerAccentPrivate: UIColor = FXColors.Purple60
+    var layerAccentPrivateNonOpaque: UIColor = FXColors.Purple60.withAlphaComponent(0.1)
+    var layerLightGrey30: UIColor = FXColors.LightGrey30
+
+    // MARK: - Actions
+    var actionPrimary: UIColor = FXColors.Blue50
+    var actionPrimaryHover: UIColor = FXColors.Blue60
+    var actionSecondary: UIColor = FXColors.LightGrey30
+    var actionSecondaryHover: UIColor = FXColors.LightGrey40
+    var formSurfaceOff: UIColor = FXColors.LightGrey30
+    var formKnob: UIColor = FXColors.White
+    var indicatorActive: UIColor = FXColors.LightGrey50
+    var indicatorInactive: UIColor = FXColors.LightGrey30
+
+    // MARK: - Text
+    var textPrimary: UIColor = FXColors.DarkGrey90
+    var textSecondary: UIColor = FXColors.DarkGrey05
+    var textDisabled: UIColor = FXColors.DarkGrey90.withAlphaComponent(0.4)
+    var textWarning: UIColor = FXColors.Red70
+    var textAccent: UIColor = FXColors.Blue50
+    var textOnColor: UIColor = FXColors.LightGrey05
+    var textInverted: UIColor = FXColors.LightGrey05
+
+    // MARK: - Icons
+    var iconPrimary: UIColor = FXColors.DarkGrey90
+    var iconSecondary: UIColor = FXColors.DarkGrey05
+    var iconDisabled: UIColor = FXColors.DarkGrey90.withAlphaComponent(0.4)
+    var iconAction: UIColor = FXColors.Blue50
+    var iconOnColor: UIColor = FXColors.LightGrey05
+    var iconWarning: UIColor = FXColors.Red70
+    var iconAccentViolet: UIColor = FXColors.Violet60
+    var iconAccentBlue: UIColor = FXColors.Blue60
+    var iconAccentPink: UIColor = FXColors.Pink60
+    var iconAccentGreen: UIColor = FXColors.Green60
+    var iconAccentYellow: UIColor = FXColors.Yellow60
+
+    // MARK: - Border
+    var borderPrimary: UIColor = FXColors.LightGrey30
+    var borderAccent: UIColor = FXColors.Blue50
+    var borderAccentNonOpaque: UIColor = FXColors.Blue50.withAlphaComponent(0.1)
+    var borderAccentPrivate: UIColor = FXColors.Purple60
 }
